simpli_api_py/requirements.txt
aiofiles==0.5.0
aioredis==1.3.1
amqp==5.0.1
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
billiard==3.6.3.0
celery==5.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
dnspython==2.0.0
email-validator==1.1.1
fastapi==0.61.1
flower==0.9.5
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
hiredis==1.1.0
httptools==0.1.1
humanize==3.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
numpy==1.19.3
orjson==3.4.3
pandas==1.1.3
pika==1.1.0
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.8
pydantic==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-multipart==0.0.5
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
Rx==1.6.1
sentry-sdk==0.19.1
six==1.15.0
slacker==0.14.0
starlette==0.13.6
tornado==6.0.4
ujson==3.2.0
urllib3==1.25.11
uvicorn==0.11.8
uvloop==0.14.0
vine==5.0.0
wcwidth==0.2.5
websockets==8.1
zstandard==0.14.0
